+++ src/constants/Extension.ts
@@ -1,24 +1,25 @@
-export const EXTENSION_KEY = 'exportall';
-export const CONFIG_FOLDERS = 'config.folderListener';
-export const CONFIG_EXCLUDE = 'config.exclude';
-export const CONFIG_RELATIVE_EXCLUDE = 'config.relExclusion';
-export const CONFIG_INCLUDE_FOLDERS = 'config.includeFoldersToExport';
-export const CONFIG_SEMIS = 'config.semis';
-export const CONFIG_QUOTE = 'config.quote';
-export const CONFIG_MESSAGE = 'config.message';
-export const CONFIG_NAMED_EXPORTS = 'config.namedExports';
+export const EXTENSION_KEY = "exportall";
+export const CONFIG_FOLDERS = "config.folderListener";
+export const CONFIG_EXCLUDE = "config.exclude";
+export const CONFIG_RELATIVE_EXCLUDE = "config.relExclusion";
+export const CONFIG_INCLUDE_FOLDERS = "config.includeFoldersToExport";
+export const CONFIG_SEMIS = "config.semis";
+export const CONFIG_QUOTE = "config.quote";
+export const CONFIG_MESSAGE = "config.message";
+export const CONFIG_NAMED_EXPORTS = "config.namedExports";
 
 export const EXTENSION_NAME = "Barrel Generator";
 
 export const COMMAND_KEYS = {
   Generate: "generate",
+  GenerateCurrentFile: "generateCurrentFile",
   Open: "open",
   AddFolder: "addFolder",
   RemoveFolder: "removeFolder",
   ExcludeFolder: "excludeFolder",
   ExcludeFile: "excludeFile",
   Include: "includeFolderFile",
-  RefreshView: "refreshView"
+  RefreshView: "refreshView",
 };
 
 export const getSettingName = (name: string) => {